Renamed Xrv_examples to Xrv_types
