+++ sympy/stats/frv_types.py
